+++ src/bit_vectors.rs
@@ -24,7 +24,7 @@
 //! | --- | :-: | :-: | :-: | :-: | :-: |
 //! | [`BitVector`] | $`O(1)`$  | $`O(u)`$ | $`O(u)`$ | $`O(1)`$ | $`u`$ |
 //! | [`BitVector<rank9sel::inner::Rank9SelIndex>`] | $`O(1)`$ | $`O(1)`$ | $`O(\lg u)`$ | -- | $`u + o(u)`$ |
-//! | [`DArray`] | $`O(1)`$ | $`O(1)`$ | $`O(1)`$ | -- | $`u + o(u)`$ |
+//! | [`BitVector<darray::inner::DArrayFullIndex>`] | $`O(1)`$ | $`O(1)`$ | $`O(1)`$ | -- | $`u + o(u)`$ |
 //!
 //! ## Plain bit vectors without index
 //!
@@ -35,15 +35,15 @@
 //!
 //! ## Plain bit vectors with index
 //!
-//! [`BitVector<rank9sel::inner::Rank9SelIndex>`] and [`DArray`] are index structures for faster queries built on [`BitVector`].
+//! [`BitVector<rank9sel::inner::Rank9SelIndex>`] and [`BitVector<darray::inner::DArrayFullIndex>`] are index structures for faster queries built on [`BitVector`].
 //!
 //! [`BitVector<rank9sel::inner::Rank9SelIndex>`] is an implementation of Vigna's Rank9 and hinted selection techniques, supporting
 //! constant-time Rank and logarithmic-time Select queries.
 //!
-//! [`DArray`] is a constant-time Select data structure by Okanohara and Sadakane.
+//! [`BitVector<darray::inner::DArrayFullIndex>`] implements the dense array technique of Okanohara and Sadakane.
 //! If you need only Select queries on dense sets (i.e., $`n/u \approx 0.5`$), this will be the most candidate.
 //! Rank/Predecessor/Successor queries are optionally enabled using the [`Rank9SelIndex`](rank9sel::inner::Rank9SelIndex) index.
-//! [`DArray`] outperforms [`BitVector<rank9sel::inner::Rank9SelIndex>`] in complexity, but the practical space overhead of [`DArray`] can be larger.
+//! This structure outperforms [`BitVector<rank9sel::inner::Rank9SelIndex>`] in complexity, but the practical space overhead can be larger.
 //!
 //!
 //! # Examples
@@ -82,7 +82,6 @@ pub mod data;
 pub mod prelude;
 pub mod rank9sel;
 
-pub use darray::DArray;
 pub use data::{BitVector, BitVectorData, BitVectorIndex, IndexBuilder, NoIndex};
 
 use anyhow::Result;