+++ src/bit_vectors/darray/mod.rs
@@ -1,0 +1,5 @@
+//! Index types implementing the dense array technique.
+
+pub mod inner;
+
+pub use inner::{DArrayFullIndex, DArrayFullIndexBuilder, DArrayIndex, DArrayIndexBuilder};